migrated_azure_fn/request_creator/requirements.txt
azure-functions
azure-identity
azure-cosmos
azure-eventgrid
